requirements.txt

torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
tqdm>=4.66.0
pyyaml>=6.0
matplotlib>=3.7.0
